Counting Char&Lines&Words from cin.get()
This program clearly illustrates the usage of "cin.get()", which reads the input string one char by one char (because it has to be assigned to a char variable). 
Personally I think the core code lies among the three if-clauses, which tells how to recognize a character ,a word and a new line.